Conflicts:
	common/coroutines.cpp
	common/coroutines.h
	devtools/create_project/msbuild.cpp
	devtools/create_project/visualstudio.cpp

+++ engines/toon/drew.h
@@ -35,7 +35,7 @@ public:
 	virtual ~CharacterDrew();
 	bool setupPalette();
 	void playStandingAnim();
-	void setPosition(int32 x, int32 y);
+	void setPosition(int16 x, int16 y);
 	void resetScale();
 	void update(int32 timeIncrement);
 	void playWalkAnim(int32 start, int32 end);
